added everything

+++ WeatherApi/Models/WeatherAPI/CurrentWeather.cs
@@ -13,10 +13,10 @@ namespace WeatherApi.Models.WeatherAPI
         public int pressure { get; set; }
         public int humidity { get; set; }
         public double dew_point { get; set; }
-        public int uvi { get; set; }
+        public decimal uvi { get; set; }
         public int clouds { get; set; }
         public int visibility { get; set; }
-        public int wind_speed { get; set; }
+        public decimal wind_speed { get; set; }
         public int wind_deg { get; set; }
         public List<Weather> weather { get; set; }
     }
@@ -35,12 +35,12 @@ namespace WeatherApi.Models.WeatherAPI
         public int pressure { get; set; }
         public int humidity { get; set; }
         public double dew_point { get; set; }
-        public double wind_speed { get; set; }
+        public decimal wind_speed { get; set; }
         public int wind_deg { get; set; }
         public double wind_gust { get; set; }
         public List<Weather> weather { get; set; }
         public int clouds { get; set; }
-        public int pop { get; set; }
+        public double pop { get; set; }
         public double uvi { get; set; }
         public double? rain { get; set; }
     }
